+++ bittytax/parsers/bittrex.py
@@ -6,10 +6,34 @@ from ..parser import DataParser
 
 WALLET = "Bittrex"
 
+def parse_bittrex_trades2(in_row):
+    if in_row[3] == "LIMIT_BUY":
+        return TransactionRecord(TransactionRecord.TYPE_TRADE,
+                                 DataParser.parse_timestamp(in_row[14]),
+                                 buy_quantity=in_row[5],
+                                 buy_asset=in_row[1].split('-')[1],
+                                 sell_quantity=in_row[8],
+                                 sell_asset=in_row[1].split('-')[0],
+                                 fee_quantity=in_row[7],
+                                 fee_asset=in_row[1].split('-')[0],
+                                 wallet=WALLET)
+    elif in_row[3] == "LIMIT_SELL":
+        return TransactionRecord(TransactionRecord.TYPE_TRADE,
+                                 DataParser.parse_timestamp(in_row[14]),
+                                 buy_quantity=in_row[8],
+                                 buy_asset=in_row[1].split('-')[0],
+                                 sell_quantity=in_row[5],
+                                 sell_asset=in_row[1].split('-')[1],
+                                 fee_quantity=in_row[7],
+                                 fee_asset=in_row[1].split('-')[0],
+                                 wallet=WALLET)
+    else:
+        raise ValueError("Unrecognised Order Type: " + in_row[2])
+
 def parse_bittrex_trades(in_row):
     if in_row[2] == "LIMIT_BUY":
         return TransactionRecord(TransactionRecord.TYPE_TRADE,
-                                 DataParser.parse_timestamp(in_row[7]),
+                                 DataParser.parse_timestamp(in_row[8]),
                                  buy_quantity=in_row[3],
                                  buy_asset=in_row[1].split('-')[1],
                                  sell_quantity=in_row[6],
@@ -19,7 +43,7 @@ def parse_bittrex_trades(in_row):
                                  wallet=WALLET)
     elif in_row[2] == "LIMIT_SELL":
         return TransactionRecord(TransactionRecord.TYPE_TRADE,
-                                 DataParser.parse_timestamp(in_row[7]),
+                                 DataParser.parse_timestamp(in_row[8]),
                                  buy_quantity=in_row[6],
                                  buy_asset=in_row[1].split('-')[0],
                                  sell_quantity=in_row[3],
@@ -45,6 +69,13 @@ def parse_bittrex_withdrawals(in_row):
                              fee_quantity=in_row[7],
                              fee_asset=in_row[1],
                              wallet=WALLET)
+
+DataParser(DataParser.TYPE_EXCHANGE,
+           "Bittrex Trades",
+           ['Uuid', 'Exchange', 'TimeStamp', 'OrderType', 'Limit', 'Quantity', 'QuantityRemaining',
+            'Commission', 'Price', 'PricePerUnit', 'IsConditional', 'Condition', 'ConditionTarget',
+            'ImmediateOrCancel', 'Closed'],
+           row_handler=parse_bittrex_trades2)
 
 DataParser(DataParser.TYPE_EXCHANGE,
            "Bittrex Trades",